requirements.txt
python-dotenv
opencv-python
picamera2
flask